SRTReader: Skip unexpected line instead of failing
Because of the unstructured nature of SRT it is more vulnerable to
corruption than other subtitle formats. Sometimes there will be more than
one empty line between two queues in a SRT subtitle, or a queue will be
missing sequence number.

Pycaption fails immediately for such input. Instead, it should make best
effort, i. e. ignore such line and continue with the hope that the rest of
the file will be valid.