pycaption/sami.py: Fix end time of caption.
If a <sync> element has multiple <p>, All end time will be incorrect except the last one.
From now, The start times of previous captions are compared with the current element. If the
start times are different and the previuos end time is 0, the previous end time is set to
the current start time.